peekaboo: Updated Dockerfile
Thid remakes the Dockerfile for peekaboo, with a proper multi stage build.
The first container builts peekaboo and the built application is then
copied to the final container.

It also purges what packages we install in each container. This together
with other 'tricks' is used to keep the image size small.

Also the placeholder replacement and appliction start, is moved to an entrypoint script for
easier maintainability.

Lastly, the peekaboo config is changed to remove the pid-file, which is
not strictly needed in the compose pipeline and the script installer.