pipeline: Updated compose.yaml and needed Files
The updates to the compose file include:
cortex version bumped to a stable release build and elasticsearch
version is updated accordingly. elasticsearch also has some new
environment varaibles to properly work in this pipeline scenario.

cortex now also has a healthcheck in compose, which is used to make
sure that services, dependent on cortex, only start when cortex is
functional (healthy)

The mariadb image is switched the official mariadb docker image and
has a healtcheck, similiar to cortex

Peekaboo gets a build context added to it can be automatically build
when running docker-compose up

The other files are the cortex configuration, which is then mounted
into the cortex container, and the peekaboo ruleset, which will be
mounted into the peekaboo container