cortex: Added set-up container
This container contains a scipt, which goes through the steps to set up
a fresh cortex installation. This is normally done manually via the web
interface. After creating the needed users in cortex, it manually places
the specified API key inside of elasticsearch, instead of letting cortex
create a random one. The last step is enabling the FileInfo Analyzer.

It then exits with code 0